+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Version 0.8.4 (Alpha)
 -   Update: The new Version of the Gravatar helper Class ([Version 0.2.0](https://github.com/SamBrishes/FoxCMS/tree/helpers/helpers)).
 -   Update: The new Version of the Hash helper class (Marked as deprecated!).
 -   Update: The new Version of the Kses helper class ([Version 0.3.0](https://pear.php.net/package/Kses)).
+-   Update: The new Version of the Pagination helper class ([Version 0.2.0](https://github.com/SamBrishes/FoxCMS/tree/helpers/pagination)).
 -   Update: The new Version of the Validate helper Class ([Version 0.2.0](https://github.com/SamBrishes/FoxCMS/tree/helpers/validate)).
 -   Update: The new Version of the ZIP helper Class ([Version 0.2.1](https://github.com/SamBrishes/FoxCMS/tree/helpers/zip)).
 -   Update: Extract each class of the `Framework.php` file into a single file due to the new format standard.